requirements.txt
pycodestyle
autopep8
telethon==1.36.0
requests==2.32.3
pandas==2.2.2
tqdm==4.66.4
SQLAlchemy==2.0.31
psycopg2-binary==2.9.3
google-api-python-client==2.137.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
openpyxl==3.1.5
oauth2client==4.1.3
gspread==6.1.2